Update to Exiled 9.6.0 (7.2.0)
