--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,33 +1,0 @@
----
-name: Bug report
-about: Create a report to help us improve
-title: ''
-labels: ''
-assignees: joker-119
-
----
-
-**Describe the bug**
-A clear and concise description of what the bug is.
-
-**To Reproduce**
-Steps to reproduce the behavior:
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
-
-**Expected behavior**
-A clear and concise description of what you expected to happen.
-
-**Server logs**
-Please include a pastebin of your localadmin log file (or both MA_log and SCP_log files if you use MultiAdmin) from the time in which the bug occured
-
-**EXILED Version ("latest" is not a version):**
-
-
-**Results of `show plugins` command in console:**
-
-
-**Additional context**
-Add any other context about the problem here.
