--- CustomItems.sln
@@ -1,16 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomItems", "CustomItems\CustomItems.csproj", "{68FA2E9D-7C31-4906-8E65-075F119A29D9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{68FA2E9D-7C31-4906-8E65-075F119A29D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{68FA2E9D-7C31-4906-8E65-075F119A29D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{68FA2E9D-7C31-4906-8E65-075F119A29D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{68FA2E9D-7C31-4906-8E65-075F119A29D9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
