--- CustomItems/Directory.Build.props
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project>
-    <!-- StyleCop Analyzers configuration -->
-    <PropertyGroup>
-        <CodeAnalysisRuleSet>$(MSBuildThisFileDirectory)\CustomItems.ruleset</CodeAnalysisRuleSet>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <AdditionalFiles Include="$(MSBuildThisFileDirectory)\stylecop.json"/>
-    </ItemGroup>
-</Project>
