--- CustomItems/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EXILED" version="6.0.0-beta.26" targetFramework="net48" />
-  <package id="Lib.Harmony" version="2.2.2" targetFramework="net472" />
-  <package id="StyleCop.Analyzers" version="1.1.118" targetFramework="net472" developmentDependency="true" />
-  <package id="YamlDotNet" version="12.0.2" targetFramework="net48" />
-</packages>
