+++ CustomItemsReborn/CustomItemsReborn.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35527.113 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomItemsReborn", "CustomItemsReborn.csproj", "{FF543006-7C57-4B99-B17E-5FF8ED369352}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FF543006-7C57-4B99-B17E-5FF8ED369352}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FF543006-7C57-4B99-B17E-5FF8ED369352}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FF543006-7C57-4B99-B17E-5FF8ED369352}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FF543006-7C57-4B99-B17E-5FF8ED369352}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
