+++ CustomItemsReborn/Items/AntiMemeticPills.cs
