+++ CustomItemsReborn/Items/LethalInjection.cs
