+++ CustomItemsReborn/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ExMod.Exiled" version="9.6.0-beta8" targetFramework="net481" />
+  <package id="Lib.Harmony" version="2.3.6" targetFramework="net481" />
+  <package id="StyleCop.Analyzers" version="1.1.118" targetFramework="net481" developmentDependency="true" />
+</packages>
